requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.1
argcomplete==1.10.3
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.8.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==3.1.0
chromadb==0.3.26
click==8.1.3
clickhouse-connect==0.6.5
coloredlogs==15.0.1
compressed-rtf==1.0.6
dataclasses-json==0.5.9
decorator==5.1.1
docx2txt==0.8
duckdb==0.8.1
ebcdic==1.1.1
extract-msg==0.28.7
faiss-cpu==1.7.4
fastapi==0.99.1
filelock==3.12.2
flatbuffers==23.5.26
frozenlist==1.3.3
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
hnswlib==0.7.0
httptools==0.5.0
huggingface-hub==0.16.2
humanfriendly==10.0
idna==3.4
IMAPClient==2.1.0
importlib-metadata==6.7.0
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
langchain==0.0.147
langchainplus-sdk==0.0.20
lxml==4.9.3
lz4==4.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.0
olefile==0.46
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pandoc==2.3
pdfminer.six==20191110
Pillow==9.5.0
plumbum==1.8.2
ply==3.11
posthog==3.0.1
protobuf==4.23.3
pulsar-client==3.2.0
pyarrow==12.0.1
pycryptodome==3.18.0
pydantic==1.10.11
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pypandoc==1.11
PyPDF2==3.0.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-pptx==0.6.21
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
referencing==0.29.1
regex==2023.6.3
requests==2.31.0
rich==13.4.2
rpds-py==0.8.8
six==1.12.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SpeechRecognition==3.8.1
SQLAlchemy==1.4.49
starlette==0.27.0
streamlit==1.24.0
sympy==1.12
tenacity==8.2.2
textract==1.6.5
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
validators==0.20.0
watchdog==3.0.0
watchfiles==0.19.0
websockets==11.0.3
xlrd==1.2.0
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.15.0
zstandard==0.21.0
